[media] Add sub-second resolution for the duration
The duration in ClutterMedia is currently expressed in integer multiples
of a second. This does not offer enough granularity for media playback
that has fractionary durations -- 2.3 seconds; 1 minute, 23.4 seconds;
1 hour, 23 minutes, 4.5 seconds; etc.

The duration value should be expressed in seconds with a sub-second
granularity; for this reason, a double should be used insted of an
unsigned integer.

The semantics haven't changed: the :duration property still exposes
seconds.

+++ clutter/clutter-media.h
@@ -72,7 +72,7 @@ void     clutter_media_set_audio_volume (ClutterMedia *media,
 gdouble  clutter_media_get_audio_volume (ClutterMedia *media);
 gboolean clutter_media_get_can_seek     (ClutterMedia *media);
 gdouble  clutter_media_get_buffer_fill  (ClutterMedia *media);
-guint    clutter_media_get_duration     (ClutterMedia *media);
+gdouble  clutter_media_get_duration     (ClutterMedia *media);
 
 G_END_DECLS
